backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
paypalrestsdk==1.2.1
python-decouple==3.8
django-crispy-forms==2.1
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
sendgrid==6.10.0
twilio==8.10.0
requests==2.31.0
django-admin-interface==0.28.6
django-colorfield==0.10.1
gunicorn==21.2.0
whitenoise==6.6.0
django-widget-tweaks==1.5.0
cryptography==41.0.8
